Changed structure and added double quotes
