+++ src/ordinalsbot.ts
@@ -1,126 +1,62 @@
-import axios, { AxiosInstance } from "axios";
-import { OrdinalsBotError } from "./OrdinalsBotError";
-import { OrdinalsBotEnv } from "./types";
-import {
-  OrdinalsBotPriceRequest,
-  OrdinalsBotPriceResponse,
-  OrdinalsBotOrderRequest,
-  OrdinalsBotOrder,
-  OrdinalsBotCollectionCreateRequest,
-  OrdinalsBotCollectionCreateResponse,
-  OrdinalsBotCollectionOrderRequest,
-  OrdinalsBotTextOrderRequest,
-  OrdinalsBotInventoryResponse,
-  OrdinalsBotReferralRequest,
-  OrdinalsBotReferralSetResponse,
-  OrdinalsBotReferralStatusResponse,
-} from "./types/v1";
+import { OrdinalsBotClient } from "./client";
+import { OrdinalsBotEnv, v1 } from "./types";
 
-const version = require("../package.json")?.version || "local";
-const packageVersion = `npm-ordinalsbot-v${version}`;
-
-export class OrdinalsBotClient {
-  public env: OrdinalsBotEnv;
-
-  private api_key: string;
-  private instanceV1: AxiosInstance;
+export class OrdinalsBot {
+  instance!: OrdinalsBotClient;
 
   constructor(key: string = "", environment: OrdinalsBotEnv = "live") {
-    this.api_key = key;
-    this.env = environment;
-
-    const createInstance = (): AxiosInstance => {
-      const client = axios.create({
-        baseURL:
-          environment === "live"
-            ? `https://api.ordinalsbot.com`
-            : `https://signet.ordinalsbot.com/api`,
-        timeout: 30000,
-        headers: {
-          "x-api-key": this.api_key,
-          Connection: "Keep-Alive",
-          "Content-Type": "application/json",
-          "Keep-Alive": "timeout=10",
-          "User-Agent": packageVersion,
-        },
-      });
-
-      client.interceptors.response.use(
-        // normalize responses
-        ({ data }) => ("data" in data ? data.data : data),
-        (err) => {
-          if (axios.isAxiosError(err)) {
-            // added to keep compatibility with previous versions
-            throw new OrdinalsBotError(
-              err.message,
-              err.response?.statusText,
-              err.response?.status
-            );
-          }
-
-          if (err instanceof Error) throw err;
-
-          return err;
-        }
-      );
-
-      return client;
-    };
-
-    this.instanceV1 = createInstance();
+    if (this.instance !== undefined) {
+      console.error("ordinalsbot.setCredentials was called multiple times");
+      return;
+    }
+    this.instance = new OrdinalsBotClient(key, environment);
   }
 
-  async getPrice(
-    priceRequest: OrdinalsBotPriceRequest
-  ): Promise<OrdinalsBotPriceResponse> {
-    return this.instanceV1.get(`/order`, {
-      params: priceRequest,
-    });
+  getPrice(
+    priceRequest: v1.OrdinalsBotPriceRequest
+  ): Promise<v1.OrdinalsBotPriceResponse> {
+    return this.instance.getPrice(priceRequest);
   }
 
-  async createOrder(order: OrdinalsBotOrderRequest): Promise<OrdinalsBotOrder> {
-    return this.instanceV1.post(`/order`, order);
+  createOrder(order: v1.OrdinalsBotOrderRequest): Promise<v1.OrdinalsBotOrder> {
+    return this.instance.createOrder(order);
   }
 
-  async getOrder(id: string): Promise<OrdinalsBotOrder> {
-    return this.instanceV1.get(`/order`, {
-      params: { id },
-    });
+  getOrder(id: string): Promise<v1.OrdinalsBotOrder> {
+    return this.instance.getOrder(id);
   }
 
-  async createCollection(
-    collection: OrdinalsBotCollectionCreateRequest
-  ): Promise<OrdinalsBotCollectionCreateResponse> {
-    return this.instanceV1.post(`/collection-create`, collection);
+  createCollection(
+    collection: v1.OrdinalsBotCollectionCreateRequest
+  ): Promise<v1.OrdinalsBotCollectionCreateResponse> {
+    return this.instance.createCollection(collection);
   }
 
-  async createCollectionOrder(
-    collectionOrder: OrdinalsBotCollectionOrderRequest
-  ): Promise<OrdinalsBotOrder> {
-    return this.instanceV1.post(`/collection-order`, collectionOrder);
+  createCollectionOrder(
+    collectionOrder: v1.OrdinalsBotCollectionOrderRequest
+  ): Promise<v1.OrdinalsBotOrder> {
+    return this.instance.createCollectionOrder(collectionOrder);
   }
 
-  async createTextOrder(
-    order: OrdinalsBotTextOrderRequest
-  ): Promise<OrdinalsBotOrder> {
-    return this.instanceV1.post(`/text-order`, order);
+  createTextOrder(
+    order: v1.OrdinalsBotTextOrderRequest
+  ): Promise<v1.OrdinalsBotOrder> {
+    return this.instance.createTextOrder(order);
   }
 
-  async getInventory(): Promise<OrdinalsBotInventoryResponse[]> {
-    return this.instanceV1.get(`/inventory`);
+  getInventory(): Promise<v1.OrdinalsBotInventoryResponse[]> {
+    return this.instance.getInventory();
   }
 
-  async setReferralCode(
-    referral: OrdinalsBotReferralRequest
-  ): Promise<OrdinalsBotReferralSetResponse> {
-    return this.instanceV1.post(`/referrals`, referral);
+  setReferralCode(
+    referral: v1.OrdinalsBotReferralRequest
+  ): Promise<v1.OrdinalsBotReferralSetResponse> {
+    return this.instance.setReferralCode(referral);
   }
 
-  async getReferralStatus(
-    referral: OrdinalsBotReferralRequest
-  ): Promise<OrdinalsBotReferralStatusResponse> {
-    return this.instanceV1.get(`/referrals`, {
-      params: referral,
-    });
+  getReferralStatus(
+    referral: v1.OrdinalsBotReferralRequest
+  ): Promise<v1.OrdinalsBotReferralStatusResponse> {
+    return this.instance.getReferralStatus(referral);
   }
 }
